+++ viteclientts/src/pages/LoginPage.tsx
@@ -3,12 +3,14 @@ import { TextField, Typography, Box, Button } from '@mui/material';
 import axios from 'axios';
 import { useCookies } from 'react-cookie';
 import { useNavigate } from 'react-router-dom';
+import { useUserStore } from '../utils/store';
 
 const LoginForm: React.FC = () => {
   const [username, setUsername] = useState<string>('');
   const [password, setPassword] = useState<string>('');
   const [errorMsg, setErrorMsg] = useState<string>('');
-  const [, setCookie] = useCookies(['session', 'username', 'role']);
+  const { setRole } = useUserStore();
+  const [, setCookie] = useCookies(['session', 'username']);
   const navigate = useNavigate();
   const isSecureCookie = import.meta.env.VITE_SECURE_COOKIE === 'true';
   const apiUrl = import.meta.env.VITE_API_URL;
@@ -26,7 +28,7 @@ const LoginForm: React.FC = () => {
       if (response.data.status === 'success') {
         setCookie('session', response.data.session, { path: '/', secure: isSecureCookie });
         setCookie('username', username, { path: '/', secure: isSecureCookie });
-        setCookie('role', response.data.role, { path: '/', secure: isSecureCookie });
+        setRole(response.data.role);
         navigate('/home');
       } else {
         setErrorMsg(response.data.message);