+++ viteclientts/src/utils/store.tsx
@@ -1,0 +1,7 @@
+import { create } from 'zustand';
+import { UserState } from '../types/types';
+
+export const useUserStore = create<UserState>(set => ({
+    role: null,
+    setRole: (role) => set({ role })
+}))